apps/life_expectancy/requirements.txt
matplotlib
mapclassify
streamlit
geopandas
plotly